refactor(ic-certified-assets): extract batch.rs from state_machine
Move the batch/chunk data types (Chunk, Batch, ComputationStatus,
CommitBatchProgress, BATCH_EXPIRY_NANOS) and the State methods that drive
batches (create_batch, create_chunks, commit_batch, delete_batch, plus the
private helpers check_batch_limits, compute_last_chunk_data,
certify_404_if_required, create_chunks_helper) into a new top-level
batch.rs.

commit_batch alone is ~250 lines of state-machine logic that doesn't share
data structures or invariants with the rest of State; collocating it with
the small helpers that only it needs gives the bulk of state_machine/mod.rs
a more uniform purpose. The remaining file is now ~580 lines focused on
asset CRUD, permissions, http_request, and state-hash bookkeeping.

State fields used by batch.rs across the module boundary are bumped to
pub(crate); same for State::dependent_keys and State::complete_set_asset_content,
both called by the commit_batch state machine.
